Fix Issue #2301 - Enable PSR-2 coding standards for Friendica addons
Convert all the PHP files formatting to the PSR-2 standards. No change
has been made to the functionalities, and any change in behavior is not
expected.

The files have been converted using util/php-cs-fixer.phar and the
config
file is .php_cs. The project-wide conversion can be run again with the
following command:

> php util/php-cs-fixer.phar fix .

Most of the changes are whitespace conversion from tabs to spaces.

Fix #2301 for Friendica addons files.