No null here:
@Hypolite
This is caused in include/conversation.php line 724. I don't know why there is
a wrong Addon::callHooks call here, but we should fix it over there instead of
suppressing the error message in the addon when adding a correct type hint.

It's good actually, type hints are allowing another level of debug.

Signed-off-by: Roland Häder <roland@mxchange.org>