nitter addon added
Treansfering the addon from the separate repository into the projects addon collection.

This addon is meant to replace links to Twitter in the body of postings with a nitter [1] instance of the admins choice. It is not meant to change links to contact prociles.

In addition to the change, a note about the changed links is placed at the bottom of the posting.

1: https://github.com/zedeus/nitter