Audon Audio Chat based on the WebRTC Addon
Audon is an audio chat function. Users can log in via their Fediverse handle, create an audio chat room and communicate with each other. 
This addon uses the existing WebRTC addon and integrates Audon as a service instead. This is a proposal to integrate further services of the Fediverse. 
The translation was only adapted for DE.